requirements.txt
streamlit>=1.18.1
plotly>=5.10.0
python-dotenv>=0.20.0
pandas>=1.4.2
sqlalchemy==1.4.37
psycopg2-binary==2.9.3